feat[nested-props]: Rewrite WriteForm story the svelte way and use nested props logic
